Добавлены тесты по client-info и content
