+++ src/test/java/Content/VersionControlTest/Android1542Test.java
@@ -1,26 +1,26 @@
-package Content.VersionControl;
+package Content.VersionControlTest;
 
 import Content.Helper.ContentRequest;
-import Content.Parameters.OnlyAndroidPrerollLight;
-import Integration.ActivatorOnlyAndroid;
+import Content.Parameters.GhostLightAndroid1542;
+import Content.Helper.FindFlightIdInActivator;
 import org.json.JSONException;
 import org.testng.annotations.BeforeTest;
 import org.testng.annotations.Test;
 
 import static org.junit.Assert.assertEquals;
 
-public class OnlyAndroidTest {
+public class Android1542Test {
     String content, idFlight;
 
     @BeforeTest
     public void receiveData() throws JSONException {
-        String body = new OnlyAndroidPrerollLight().getBody().toString();
-        idFlight = new ActivatorOnlyAndroid().getFlightId();
+        String body = new GhostLightAndroid1542().getBody().toString();
+        idFlight = new FindFlightIdInActivator().Android15421();
         content = new ContentRequest(body).getContent();
     }
 
     @Test
-    public void OnlyAndroid() {
+    public void Android1542() {
         assertEquals(idFlight, content);
         //System.out.println(content + " " + idFlight);
     }
