+++ src/test/java/Content/VersionControlTest/AndroidIOSVersionsTest.java
@@ -1,23 +1,23 @@
-package Content.VersionControl;
+package Content.VersionControlTest;
 
 import Content.Helper.ContentRequest;
-import Content.Parameters.GhostLightAndroid;
-import Content.Parameters.GhostLightIOS;
-import Integration.ActivatorAndroidIOS;
+import Content.Parameters.GhostDarkAndroid;
+import Content.Parameters.GhostDarkIOS;
+import Content.Helper.FindFlightIdInActivator;
 import org.json.JSONException;
 import org.testng.annotations.BeforeTest;
 import org.testng.annotations.Test;
 
 import static org.junit.Assert.assertEquals;
 
-public class AndroidIOSTest {
+public class AndroidIOSVersionsTest {
     String idFlight, content1, content2;
 
     @BeforeTest
     public void receiveData() throws JSONException {
-        String body1 = new GhostLightIOS().getBody().toString();
-        String body2 = new GhostLightAndroid().getBody().toString();
-        idFlight = new ActivatorAndroidIOS().getFlightId();
+        String body1 = new GhostDarkIOS().getBody().toString();
+        String body2 = new GhostDarkAndroid().getBody().toString();
+        idFlight = new FindFlightIdInActivator().IOSAndroidVersions();
         content1 = new ContentRequest(body1).IOSVersions(body1);
         content2 = new ContentRequest(body2).AndroidVersions(body2);
     }
