+++ src/test/java/Content/VersionControlTest/NotOnlyAndroidTest.java
@@ -1,27 +1,27 @@
-package Content.VersionControl;
+package Content.VersionControlTest;
 
 import Content.Helper.ContentRequest;
 import Content.Parameters.OnlyIOSPrerollLight;
-import Integration.ActivatorOnlyIOS;
+import Content.Helper.FindFlightIdInActivator;
 import org.json.JSONException;
 import org.testng.annotations.BeforeTest;
 import org.testng.annotations.Test;
 
-import static org.junit.Assert.assertEquals;
+import static org.junit.Assert.assertNotEquals;
 
-public class OnlyIOSTest {
+public class NotOnlyAndroidTest {
     String content, idFlight;
 
     @BeforeTest
     public void receiveData() throws JSONException {
         String body = new OnlyIOSPrerollLight().getBody().toString();
-        idFlight = new ActivatorOnlyIOS().getFlightId();
+        idFlight = new FindFlightIdInActivator().OnlyAndroid();
         content = new ContentRequest(body).getContent();
     }
 
     @Test
-    public void OnlyIOS() {
-        assertEquals(idFlight, content);
+    public void NotOnlyAndroid() {
+        assertNotEquals(idFlight, content);
         //System.out.println(content + " " + idFlight);
     }
 }
