+++ src/test/java/Content/VersionControlTest/NotOnlyIOSTest.java
@@ -1,26 +1,26 @@
-package Content.VersionControl;
+package Content.VersionControlTest;
 
 import Content.Helper.ContentRequest;
-import Content.Parameters.OnlyIOSPrerollLight;
-import Integration.ActivatorOnlyAndroid;
+import Content.Parameters.OnlyAndroidPrerollLight;
+import Content.Helper.FindFlightIdInActivator;
 import org.json.JSONException;
 import org.testng.annotations.BeforeTest;
 import org.testng.annotations.Test;
 
 import static org.junit.Assert.assertNotEquals;
 
-public class NotOnlyAndroidTest {
+public class NotOnlyIOSTest {
     String content, idFlight;
 
     @BeforeTest
     public void receiveData() throws JSONException {
-        String body = new OnlyIOSPrerollLight().getBody().toString();
-        idFlight = new ActivatorOnlyAndroid().getFlightId();
+        String body = new OnlyAndroidPrerollLight().getBody().toString();
+        idFlight = new FindFlightIdInActivator().OnlyIOS();
         content = new ContentRequest(body).getContent();
     }
 
     @Test
-    public void NotOnlyAndroid() {
+    public void NotOnlyIOS() {
         assertNotEquals(idFlight, content);
         //System.out.println(content + " " + idFlight);
     }
